Add support of mu site while removing plugin data on uninstall
